scx_stats,scx_lavd: describe non-nested stats meta
In cases where non-nested structs are used for statistics (such as in scx_lavd), the describe_metadata method cannot display metadata for all of them.
This is because the current recursion mechanism does not directly support multiple monitoring modes that use different, unrelated structs.

In scx_lavd, two distinct monitoring modes rely on separate structs. Therefore, it was necessary to create a wrapper to utilize the existing describe_metadata method in a way that accommodates both these separate statistics.
This makes it possible to display the metadata description either selectively or for n statistics simultaneously.

The issue is reported in sched-ext#783

Signed-off-by: Luigi De Matteis <ldematteis123@gmail.com>